+++ src/client.ts
@@ -20,6 +20,7 @@ import {createCommandLogs} from "#commands/CommandLogs/CommandLogs";
 import {createCommandStop} from "#commands/CommandStop/CommandStop";
 import {createCommandRestart} from "#commands/CommandRestart/CommandRestart";
 import {createCommandUpdate} from "#commands/CommandUpdate/CommandUpdate";
+import {createCommandIdentity} from "#commands/CommandIdentity/CommandIdentity";
 import {ApiClient} from "#api-client/ApiClient";
 import {readPackageJson} from "#package-json/packageJson";
 
@@ -65,6 +66,7 @@ program.addCommand(createCommandLogs(ctx, program))
 program.addCommand(createCommandStop(ctx, program))
 program.addCommand(createCommandRestart(ctx, program))
 program.addCommand(createCommandUpdate(ctx, program))
+program.addCommand(createCommandIdentity(ctx, program))
 
 program.action(({options}: ParseResult) => {
   if (options.version) {